+++ acme/tests/client_test.py
@@ -5,12 +5,9 @@ import datetime
 import http.client as http_client
 import json
 import unittest
+from unittest import mock
 
 import josepy as jose
-try:
-    import mock
-except ImportError: # pragma: no cover
-    from unittest import mock # type: ignore
 import OpenSSL
 import requests