+++ certbot/certbot/tests/util.py
@@ -8,6 +8,7 @@ import shutil
 import sys
 import tempfile
 import unittest
+import warnings
 
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
@@ -27,6 +28,12 @@ from certbot.display import util as display_util
 
 try:
     import mock
+    warnings.warn(
+        "The external mock module is being used for backwards compatibility "
+        "since it is available, however, future versions of Certbot's tests will "
+        "use unittest.mock. Be sure to update your code accordingly.",
+        PendingDeprecationWarning
+    )
 except ImportError: # pragma: no cover
     from unittest import mock # type: ignore